Add chef filtering to map
Map can measure distance between a chef and user
and populate map with chefs near the user.
Refactored code to create 3 separate files for
geolocation: a map component, the file that
gets the geolocation, and a file that makes
axios call and filters chefs by distance.
